LayoutTests/media/video-error-abort-expected.txt

EXPECTED (video.error == 'null') OK
EVENT(loadstart)
EVENT(abort)
EXPECTED (video.error.code == '1') OK
EVENT(loadstart)
EVENT(load)
EXPECTED (video.error == 'null') OK
END OF TEST
